chore: split core chunk to transform & shared
