+++ packages/core/src/index.ts
@@ -1,14 +1,5 @@
 export * from './colors/colorbrewer'
-
 export * from './colors/uno'
 export * from './colors/w3cx11'
-export * from './transform/hex'
-export * from './transform/hsb'
-export * from './transform/hsl'
-export * from './transform/keywords'
 
-export * from './transform/lab'
-export * from './transform/rgb'
 export * from './types'
-
-export * from './utils'
