+++ packages/magic-color/src/core/index.ts
@@ -1,3 +1,2 @@
 export * from './basic'
 export * from './mc'
-export * from './types'
