+++ packages/magic-color/src/core/mc.ts
@@ -2,8 +2,8 @@ import type { Colors, ColorType, Opacity } from '@magic-color/core'
 import { calcAPCA, calcWCAG, getReadableTextColor, isWarmColor, reverseAPCA } from '../contrast'
 import { deltaE } from '../delta-e'
 import { hash } from '../hash'
+import { random } from '../random'
 import { getColorName, theme } from '../theme'
-import { random } from '../utils'
 import { Magicolor } from './basic'
 import { SupportTypes, valid } from './utils'
