+++ packages/magic-color/src/index.ts
@@ -2,6 +2,6 @@ export * from './contrast'
 
 export * from './core'
 export * from './hash'
+export * from './random'
 export * from './theme'
-export * from './utils'
 export * from '@magic-color/core'
