+++ packages/magic-color/src/random.ts
