+++ packages/shared/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "magic-color",
+  "name": "@magic-color/shared",
   "type": "module",
   "version": "1.3.0",
-  "description": "Magic color creator.",
+  "description": "Magic color common utils library",
   "author": "Chris <hizyyv@gmail.com>",
   "license": "MIT",
   "homepage": "https://github.com/zyyv/magic-color#readme",
@@ -40,10 +40,5 @@
   "scripts": {
     "build": "unbuild",
     "dev": "unbuild --stub"
-  },
-  "dependencies": {
-    "@magic-color/core": "workspace:*",
-    "@magic-color/shared": "workspace:*",
-    "@magic-color/transform": "workspace:*"
   }
 }
