+++ packages/transform/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "magic-color",
+  "name": "@magic-color/transform",
   "type": "module",
   "version": "1.3.0",
-  "description": "Magic color creator.",
+  "description": "Magic color transform library",
   "author": "Chris <hizyyv@gmail.com>",
   "license": "MIT",
   "homepage": "https://github.com/zyyv/magic-color#readme",
@@ -43,7 +43,6 @@
   },
   "dependencies": {
     "@magic-color/core": "workspace:*",
-    "@magic-color/shared": "workspace:*",
-    "@magic-color/transform": "workspace:*"
+    "@magic-color/shared": "workspace:*"
   }
 }
